--- src/apps/ScopeChangeRequest/Components/MenuButton/index.ts
@@ -1,3 +1,0 @@
-export * from './Components/IconMenu';
-export * from './Components/MenuButton';
-export * from './Types/menuItem';
