+++ src/components/OverlayMenu/src/components/ButtonMenu/buttonMenu.styles.ts
@@ -1,0 +1,11 @@
+import { Typography } from '@equinor/eds-core-react';
+import styled from 'styled-components';
+
+export const Wrapper = styled.div`
+    display: flex;
+    flex-direction: column;
+`;
+
+export const MenuText = styled(Typography)`
+    font-size: 16px;
+`;
