+++ src/components/OverlayMenu/src/components/IconMenu/iconMenu.styles.ts
@@ -1,0 +1,13 @@
+import { Typography } from '@equinor/eds-core-react';
+import styled from 'styled-components';
+
+export const Wrapper = styled.div`
+    display: flex;
+    flex-direction: column;
+    height: auto;
+    width: auto;
+`;
+
+export const MenuText = styled(Typography)`
+    font-size: 16px;
+`;
