+++ src/components/OverlayMenu/src/components/index.ts
@@ -1,0 +1,2 @@
+export * from './ButtonMenu/ButtonMenu';
+export * from './IconMenu/IconMenu';
