+++ src/components/OverlayMenu/src/index.ts
@@ -1,0 +1,2 @@
+export * from './types/menuItem';
+export * from './components';
