+++ vite.config.ts
@@ -41,6 +41,7 @@ export default defineConfig({
             '@equinor/GardenUtils': path.resolve(__dirname, './src/packages/GardenUtils'),
             '@equinor/echo3dweb-viewer': path.resolve(__dirname, './packages/echo3dViewer/src'),
             '@equinor/lighthouse-powerbi': path.resolve(__dirname, './src/modules/powerBI'),
+            '@equinor/overlay-menu': path.resolve(__dirname, './src/components/OverlayMenu'),
             '@equinor/lighthouse-powerbi-viewer': path.resolve(
                 __dirname,
                 './src/Core/PowerBiViewer'
